--- Project Serenity/comrade_cfg.py
@@ -1,5 +1,0 @@
-cfg = {"LETHALITY": 0.0, "kickList":{308287917556498452: 0, 292015630637596672: 0, 116275390695079945: 0, 369165245307944962: 0, 358652973330595841: 0, 471542070817849355: 0, 434101694284955649: 0, 252128902418268161: 0, 209451668557922307: 0, 134073775925886976: 0, 426761094036586498: 0, 333621621116108800: 0, 206955935229280256: 0, 328318055002406914: 0, 281603116665667586: 0, 268512516957536257: 0, 235088799074484224: 0, 423184609006583820: 0, 482664998590742528: 0, 417568038335021059: 0, 186655764406992896: 0, 329330474830790657: 0, 270904126974590976: 0, 268172436669726735: 0, 494878217153413152: 0, 233977372620095488: 0, 560260885944336384: 0, 348216094751850496: 0, 408785106942164992: 0, 244234418007441408: 0, 233032569166364672: 0, 206117782226468866: 0, 519269787738701864: 0, 200614685760946176: 0, 529135112819048449: 0, 637861879238230067: 0, 448145698190524440: 0, 464894966368567306: 0, 204397270152642561: 0, 341736321410400276: 0, 155149108183695360: 0, 66137108124803072: 0, 186638027634769920: 0, 308028236418252801: 0, 268173116474130443: 0, 370682062623080448: 0, 162646812316925953: 0, 307618081096597505: 0, 267785734323306496: 0, 178059181574586368: 0, 158730602537811971: 0, 317619178536435712: 0, 366276166228639748: 0, 417798584910479360: 0, 356116010367057931: 0, 422461687963058176: 0, 135410674443485184: 0, 294647638253371412: 0, 502128898767847425: 0, 143813149601169408: 0, 393517365846999051: 0, 231184271438643210: 0, 545672836124246024: 0, 555116957729423380: 0, 541732349134766102: 0, 524352248147476490: 0, 545006913692106753: 0, 193511451980333058: 0, 318552464356016131: 0, 393535226380288010: 0, 285913892712873985: 0, 472141928578940958: 0, 513808186940063744: 0, 176018696936882177: 0},
-"kickVotes":{308287917556498452: [], 292015630637596672: [], 116275390695079945: [], 369165245307944962: [], 358652973330595841: [], 471542070817849355: [], 434101694284955649: [], 252128902418268161: [], 209451668557922307: [], 134073775925886976: [], 426761094036586498: [], 333621621116108800: [], 206955935229280256: [], 328318055002406914: [], 281603116665667586: [], 268512516957536257: [], 235088799074484224: [], 423184609006583820: [], 482664998590742528: [], 417568038335021059: [], 186655764406992896: [], 329330474830790657: [], 270904126974590976: [], 268172436669726735: [], 494878217153413152: [], 233977372620095488: [], 560260885944336384: [], 348216094751850496: [], 408785106942164992: [], 244234418007441408: [], 233032569166364672: [], 206117782226468866: [], 519269787738701864: [], 200614685760946176: [], 529135112819048449: [], 637861879238230067: [], 448145698190524440: [], 464894966368567306: [], 204397270152642561: [], 341736321410400276: [], 155149108183695360: [], 66137108124803072: [], 186638027634769920: [], 308028236418252801: [], 268173116474130443: [], 370682062623080448: [], 162646812316925953: [], 307618081096597505: [], 267785734323306496: [], 178059181574586368: [], 158730602537811971: [], 317619178536435712: [], 366276166228639748: [], 417798584910479360: [], 356116010367057931: [], 422461687963058176: [], 135410674443485184: [], 294647638253371412: [], 502128898767847425: [], 143813149601169408: [], 393517365846999051: [], 231184271438643210: [], 545672836124246024: [], 555116957729423380: [], 541732349134766102: [], 524352248147476490: [], 545006913692106753: [], 193511451980333058: [], 318552464356016131: [], 393535226380288010: [], 285913892712873985: [], 472141928578940958: [], 513808186940063744: [], 176018696936882177: []},
-"THREATS": {268173116474130443: 228922, 545672836124246024: 1, 545006913692106753: 2}, "OPS": [66137108124803072, 178059181574586368], "KICK_REQ":6,
-"KICK_SAFE": [66137108124803072, 637861879238230067, 472141928578940958, 408785106942164992], "BANNED_WORDS": set(), "PURGE":[], "LAST_DAILY":"2019-12-15",
-"USER_BANNED_WORDS": {268173116474130443: {'kevin', 'pb', 'hentai', 'penis', 'mute', 'nae'}, 545672836124246024: {'breaking', 'just'}, 545006913692106753: set()}}
