+++ comrade_modules.py
